toolkit/components/passwordmgr/test/test_basic_form_2pw_2.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Password Manager</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>  
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Password Manager test: forms with 2 password fields (adding new logins)
<p id="display"></p>

<div id="content" style="display: none">
  <p>The next three forms are <b>user/pass/pass-new</b>, as all-empty, prefilled user(only), and prefilled user/pass</p>
  <form id="form1" onsubmit="return checkSubmit(1)" action="formtest.js" method="get">
    <input  type="text"       name="new-uname1"  value="">
    <input  type="password"   name="new-pword1"  value="">
    <input  type="password"   name="new-pword1B" value="">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

  <form id="form2" onsubmit="return checkSubmit(2)" action="formtest.js" method="get">
    <input  type="text"       name="new-uname2"  value="newuser2">
    <input  type="password"   name="new-pword2"  value="">
    <input  type="password"   name="new-pword2B" value="">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

  <form id="form3" onsubmit="return checkSubmit(3)" action="formtest.js" method="get">
    <input  type="text"       name="new-uname3"  value="newuser3">
    <input  type="password"   name="new-pword3"  value="newpass3">
    <input  type="password"   name="new-pword3B" value="">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>


  <p>The next three forms are <b>user/pass-new/pass</b>, as all-empty, prefilled user(only), and prefilled user/pass</p>
  <form id="form4" onsubmit="return checkSubmit(4)" action="formtest.js" method="get">
    <input  type="text"       name="new-uname4"  value="">
    <input  type="password"   name="new-pword4B" value="">
    <input  type="password"   name="new-pword4"  value="">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

  <form id="form5" onsubmit="return checkSubmit(5)" action="formtest.js" method="get">
    <input  type="text"       name="new-uname5"  value="newuser5">
    <input  type="password"   name="new-pword5B" value="">
    <input  type="password"   name="new-pword5"  value="">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

  <form id="form6" onsubmit="return checkSubmit(6)" action="formtest.js" method="get">
    <input  type="text"       name="new-uname6"  value="newuser6">
    <input  type="password"   name="new-pword6B" value="">
    <input  type="password"   name="new-pword6"  value="newpass6">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>
</div>

<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Password Manger: form fill, 2 password fields **/

// If a form has two password fields, other things may be going on....
//
// 1 - The user might be creating a new login (2nd field for typo checking)
// 2 - The user is changing their password (old and new password each have field)
//
// This test is for case #1.

var numSubmittedForms = 0;
var numStartingLogins = 0;

function startTest() {
  // Check for unfilled form 1
  is($_("new-uname1").value,  "", "Checking username1");
  is($_("new-pword1").value,  "", "Checking password1");
  is($_("new-pword1B").value, "", "Checking password1B");

  // Check for unfilled form 2
  is($_("new-uname2").value,  "newuser2", "Checking username2");
  is($_("new-pword2").value,  "", "Checking password2");
  is($_("new-pword2B").value, "", "Checking password2B");

  // Check for unfilled form 3
  is($_("new-uname3").value,  "newuser3", "Checking username3");
  is($_("new-pword3").value,  "newpass3", "Checking password3");
  is($_("new-pword3B").value, "", "Checking password3B");

  // Check for unfilled form 4
  is($_("new-uname4").value,  "", "Checking username4");
  is($_("new-pword4").value,  "", "Checking password4");
  is($_("new-pword4B").value, "", "Checking password4B");

  // Check for unfilled form 5
  is($_("new-uname5").value,  "newuser5", "Checking username5");
  is($_("new-pword5").value,  "", "Checking password5");
  is($_("new-pword5B").value, "", "Checking password5B");

  // Check for unfilled form 6
  is($_("new-uname6").value,  "newuser6", "Checking username6");
  is($_("new-pword6").value,  "newpass6", "Checking password6");
  is($_("new-pword6B").value, "", "Checking password6B");


  // Fill in the username and password fields, as if we are creating new accounts.
  // Form 1
  $_("new-uname1").value  = "newuser1";
  $_("new-pword1").value  = "newpass1";
  $_("new-pword1B").value = "newpass1";
  // Form 2
  //$_("new-uname2").value  = "newuser2";  (already set in form)
  $_("new-pword2").value  = "newpass2";
  $_("new-pword2B").value = "newpass2";
  // Form 3
  //$_("new-uname3").value  = "newuser3";
  //$_("new-pword3").value  = "newpass3";
  $_("new-pword3B").value = "newpass3";
  // Form 4
  $_("new-uname4").value  = "newuser4";
  $_("new-pword4").value  = "newpass4";
  $_("new-pword4B").value = "newpass4";
  // Form 5
  //$_("new-uname5").value  = "newuser5";
  $_("new-pword5").value  = "newpass5";
  $_("new-pword5B").value = "newpass5";
  // Form 6
  //$_("new-uname6").value  = "newuser6";
  //$_("new-pword6").value  = "newpass6";
  $_("new-pword6B").value = "newpass6";


  var button = getFormSubmitButton(1);
  button.click();
}


// Called by each form's onsubmit handler.
function checkSubmit(formNum) {
  numSubmittedForms++;

  // End the test at the last form.
  if (formNum == 6) {
    is(numSubmittedForms, 6, "Ensuring all forms were submitted for testing.");;

    var numEndingLogins = countLogins();

    todo(false, "Need a mechanism to auto-save submitted logins, so we can check to see if it worked.");
    if (false) {
      ok(numEndingLogins > 0, "counting logins at end");
      is(numStartingLogins, numEndingLogins + 6, "counting logins at end");
    }

    SimpleTest.finish();
    return false; // return false to cancel current form submission
  }

  // submit the next form.
  var button = getFormSubmitButton(formNum + 1);
  button.click();

  return false; // return false to cancel current form submission
}


function getFormSubmitButton(formNum) {
  var form = $("form" + formNum); // by id, not name
  ok(form != null, "getting form " + formNum);

  // we can't just call form.submit(), because that doesn't seem to invoke the form onsubmit handler.
  var button = form.firstChild;
  while (button && button.type != "submit") { button = button.nextSibling; }
  ok(button != null, "getting form submit button");

  return button;
}


// Counts the number of logins currently stored by password manager.
function countLogins() {
  netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');

  var count = 0;
  var enum = pwmgr.enumerator;
  while (enum.hasMoreElements()) { count++; enum.getNext(); }

  return count;
}

// Get the pwmgr service
netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');

var Cc_pwmgr = Components.classes["@mozilla.org/passwordmanager;1"];
ok(Cc_pwmgr != null, "Access Cc[@mozilla.org/passwordmanager;1]");

var Ci_pwmgr = Components.interfaces.nsIPasswordManager;
ok(Ci_pwmgr != null, "Access Ci.nsIPasswordManager");

var pwmgr = Cc_pwmgr.getService(Ci_pwmgr);
ok(pwmgr != null, "pwmgr getService()");


window.onload = startTest;

SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
